add go1.24-specific implementation of h2c protocol
